@@ -4,18 +4,19 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@0.1.10.
- * To regenerate, run `npx convex codegen`.
+ * Generated by convex@1.8.0.
+ * To regenerate, run `npx convex dev`.
  * @module
  */
 
 import type {
   DataModelFromSchemaDefinition,
-  DocumentMapFromSchemaDefinition,
-} from "convex/schema";
-import type { TableNamesInDataModel } from "convex/server";
-import { GenericId, GenericIdConstructor } from "convex/values";
-import schema from "../schema";
+  DocumentByName,
+  TableNamesInDataModel,
+  SystemTableNames,
+} from "convex/server";
+import type { GenericId } from "convex/values";
+import schema from "../schema.js";
 
 /**
  * The names of all of your Convex tables.
@@ -27,25 +28,26 @@ export type TableNames = TableNamesInDataModel<DataModel>;
  *
  * @typeParam TableName - A string literal type of the table name (like "users").
  */
-export type Document<TableName extends TableNames> =
-  DocumentMapFromSchemaDefinition<typeof schema>[TableName];
+export type Doc<TableName extends TableNames> = DocumentByName<
+  DataModel,
+  TableName
+>;
 
 /**
  * An identifier for a document in Convex.
  *
  * Convex documents are uniquely identified by their `Id`, which is accessible
- * on the `_id` field. To learn more, see [Data Modeling](https://docs.convex.dev/using/data-modeling).
+ * on the `_id` field. To learn more, see [Document IDs](https://docs.convex.dev/using/document-ids).
  *
  * Documents can be loaded using `db.get(id)` in query and mutation functions.
  *
- * **Important**: Use `myId.equals(otherId)` to check for equality.
- * Using `===` will not work because two different instances of `Id` can refer
- * to the same document.
+ * IDs are just strings at runtime, but this type can be used to distinguish them from other
+ * strings when type checking.
  *
  * @typeParam TableName - A string literal type of the table name (like "users").
  */
-export type Id<TableName extends TableNames> = GenericId<TableName>;
-export const Id: GenericIdConstructor<TableNames> = GenericId;
+export type Id<TableName extends TableNames | SystemTableNames> =
+  GenericId<TableName>;
 
 /**
  * A type describing your Convex data model.
