@@ -1,12 +1,9 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-
 import { ConvexProvider, ConvexReactClient } from "convex/react";
-import clientConfig from "../../convex/_generated/clientConfig";
 import Head from "next/head";
-const convex = new ConvexReactClient(clientConfig, {
-  unsavedChangesWarning: false,
-});
+
+const convex = new ConvexReactClient(process.env.NEXT_PUBLIC_CONVEX_URL!);
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
